% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_above.R
\name{vol_above}
\alias{vol_above}
\title{Calculate soil volume above a reference plane}
\usage{
vol_above(mesh, z_displ_mm = 0.01, ball_radius_mm = 0.005)
}
\arguments{
\item{mesh}{triangular mesh of class "mesh3d"}

\item{z_displ_mm}{vertical offset to prevent removal of vertices near the origin}

\item{ball_radius_mm}{radius in mm for the ball-pivoting mesh reconstruction}
}
\value{
numeric vector of length 1, representing the cm3 of soil
residing above the reference plane
}
\description{
One half of the total volume computation.
}
\details{
\lifecycle{experimental}

This applies a sort of hack to slice the mesh and keep only points
residing above the x-y plane. The "hack" portion is translating the
actual reference plane away from the mesh by a very small distance. So long
as the ball radius is smaller than this distance, one can be confident that
the reference plane will be preserved as part of the new mesh. The volume
attributable to the offset is subtracted before completing the calculation.
The meshes used in this project can \emph{never} be "watertight" and it is my
hope that so long as the mesh is 2-manifold, the volume computation is
accurate. This function calls \code{\link[Rvcg:vcgClean]{Rvcg::vcgClean()}} to remove non-manifold
faces, then remove unreferenced vertices, and (finally) remove non-manifold
vertices. This approach seems to resolve the non-manifoldness of the meshes
induced by downsampling, cylinder removal, and top/bottom removal.
}
